fix(ci): restore current main quality gates
